library_requirements.txt
clr
customtkinter
pandas
openpyxl
tk
datetime
pillow
tkwebview2
streamlit
numpy
pickles
opencv-python
shutils
face_recognition